feat: Add advanced switches and docs

+++ framework/endtoendbase.py
@@ -213,9 +213,11 @@ def cloned_repository_at_revision(url: str, version: str):
 
     try:
         run(["git", "checkout", version])
-        run(["git", "pull"])
-        run(["git", "reset", "--hard", "HEAD"])
-        run(["git", "clean", "-fx"])
+
+        if os.getenv("SKIP_GIT_PULL") != "true":
+            run(["git", "pull"])
+            run(["git", "reset", "--hard", "HEAD"])
+            run(["git", "clean", "-fx"])
         yield
     finally:
         os.chdir(pwd)
